test(input): merge render spec files
